+++ CHANGELOG.md
@@ -7,6 +7,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Added
+
+* Auto configuration of optimal chunk sizes to texture size as well as checks
+to ensure that the tiles are divisible into each other had been added.
+
 ## [0.2.2] - 2020-11-23
 
 ### Added